+++ zipline_app/forms.py
@@ -2,6 +2,7 @@ from django.forms import ModelForm
 from .models.zipline_app.order import Order
 from .models.zipline_app.fill import Fill
 from .models.zipline_app.asset import Asset
+from .models.zipline_app.account import Account
 
 class OrderForm(ModelForm):
   class Meta:
@@ -17,4 +18,9 @@ class AssetForm(ModelForm):
   class Meta:
     model = Asset
     fields = ['asset_symbol','asset_exchange','asset_name']
+
+class AccountForm(ModelForm):
+  class Meta:
+    model = Account
+    fields = ['account_symbol']